pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyptx-templar"
version = "0.1.3"
description = "Easily fill PowerPoint templates with Python"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business :: Office Suites",
    "Topic :: Software Development :: Libraries",
]
keywords = ["template", "PowerPoint", "ppt", "pptx", "office"]
dependencies = [
    "python-pptx >= 0.6",
    "Pillow >= 9"
]
requires-python = ">=3.7"

[project.urls]
Homepage = "https://github.com/pgimalac/pyptx-templar"
